Efficient batched L-BFGS updates
Summary:
Implements an efficient batched version of updating the compact representations of the L-BFGS low-rank approximations of the Hessian and its inverse. To be used in the QP solver (the inner solver of the SQP algorithm).

For a description of the updates, see Section 2.5 of the following pdf: {F141532639}

The description in the pdf is for the non-batched case, the implementation here does these updates in a batched fashion.

Note that this will throw a ValueError when computing `gamma` turns out to mean dividing by zero. This happens when  the underlying function is locally flat (e.g. if the function is linear then all the `y_k` will be zero)

Reviewed By: bletham

Differential Revision: D10361906

fbshipit-source-id: 67ef5a061fa273b8a5150733d9c417ced2a7eb54