Add KG without discretization to botorch
Summary:
As discussed, this allows for computing q-KG via optimizing over X, X_fantasies, and X_old simultaneously.

Also fixes a bug in the discretized KG where the old value
was computed using the new points.  Added an option to preserve
or remove that behavior based on an input boolean.

Also, adds in missing cost needed for multi-fidelity optimization.

Reviewed By: Balandat

Differential Revision: D10300515

fbshipit-source-id: 196043f54faf9871522502ab76942cc08acb512c