Pending arms for botorch batch acquisition functions
Summary:
This wraps the rest of the batch acquisition functions
into modules and enables them to do pending arms in the way
that I previously proposed.

In that previous notebook I also discussed how this could be
used for sequential greedy optimization.  With this diff that
would involve creating a sequence of Modules with the only
difference in X_pending.  Alternatively we could have
a setter for X_pending.

Ensures that X_pending has no gradient.

Reviewed By: Balandat

Differential Revision: D10262637

fbshipit-source-id: e7195cd9eca9a2a01f7f3b1b54e6e4f60c722faa