Add fixed_features to gen_candidates
Summary: This enables one to specify that certain features are supposed to be fixed in value according to an input dictionary during candidate generation.  If the dictionary has a None value, it just assumes that feature is meant to be fixed.  If the value is not None, it fixes the feature to that value.

Reviewed By: Balandat

Differential Revision: D10501541

fbshipit-source-id: e7f15d8f6d4b82415f1583ba94c521de34890d22