Fix shape issue in Single Task Botorch Model
Summary: This fixes a shape issue in the SingleTaskBoTorchModel due to q_batch_initialization returning a t x n x d tensor, where t is the number of t-batches and n is the number of candidate points (replaced q). Also, I added a clone() call in gen_candidates() so that we copy the candidates tensor in param_trajectory (before, updates to candidates would update every element in param_trajectory).

Reviewed By: bletham

Differential Revision: D12871235

fbshipit-source-id: 9aeb9c7481c64d41fc56763619f8d10d5c2f3435