cleanup deterministic optimization
Summary:
We had an unused seed as input to all batch
acquisition functions.  This updates batch modules to actually use a seed, and
changes the batch acquisition functions to take in a base_samples of the
appropriate size from the batch modules.  This is more efficient since we no
longer generate the same MVN draw many times.  This also lets us ensure that
each t-batch is optimizing the exact same function when using random restarts.

Reviewed By: Balandat

Differential Revision: D13280021

fbshipit-source-id: c05a24faa030c05fb7e897cfdbdd316dc7bea2a0