add batch module for q-KG without discretization
Summary: The batch acquisition function existed for q-kg without discretization but a batch module had not been created yet.  The logic for q-kg without discretization requires some subtle changes because it includes auxiliary optimization variables (X_fantasies and X_old).  We add an extract_candidates functionality to the Acquisition module that is needed to separate the actual candidates and auxiliary variables, both stored in the X tensor for this module.

Reviewed By: Balandat

Differential Revision: D13359130

fbshipit-source-id: afd374b9a5c051813e3051d129bba1896de88e81