Linear (In)Equlity constraints for scipy optimization
Summary:
Adds the ability to specify linear (non-box) constraints on the input parameters of the scipy candidate generation.
Useful e.g. for enforcing order constraints, or optimizing subject to a budget constraint in the fidelity-aware setting.

While the scipy optimizer also allows general non-linear constraints, this diff only supports linear constraints (the optimization problems are typically hard enough, we don't need to suggest people can do all kinds of crazy things with ease...).

This is botorch-only for now, and still needs to be hooked up with Lazarus.

Reviewed By: bletham

Differential Revision: D13427014

fbshipit-source-id: 282bdca4e550e7b8efac69dee4ad9d77a986af3e