Fix cuda unit tests
Summary:
These are failing for me on the GPU on master b/c of numerical issues.
E.g. `torch.tensor([1.0, 1.0, 1.0], device="cuda").mean().item()` may not be exactly 1.0, but 1e-6 close.

I am not sure whether these have been failing all along. Note that sandcastle won't pick up this kind of stuff since it doesn't run the cuda tests.
sdaulton, did you run these successfully on a devgpu?

If so this may be due to some changes deep in the cuda backend. If that's the case, let's see if that was a bad commit and whether they keep on failing going forward.

Reviewed By: bletham

Differential Revision: D13517020

fbshipit-source-id: 6cec0eb4b29215756db5ae1dfd932a944b089302