refit model after exceptions in closed loop
Summary: Refit the model after exceptions to address cases where model fitting suceeds, but an exception is raised at prediction time. This is not as clean as I would like, but this keeps any observed data such that the benchmarks reflect our ability to retry after exceptions without re-evaluating in a real optimization scenario.

Differential Revision: D13560964

fbshipit-source-id: 5ab9bab50dbe9741e53b2162b1bc7aa9effc0342