refit model after exceptions in closed loop
Summary: Refit the model after exceptions to address cases where model fitting suceeds, but an exception is raised at prediction time. This is not as clean as I would like, but this keeps any observed data such that the benchmarks reflect our ability to retry after exceptions without re-evaluating in a real optimization scenario.

Reviewed By: Balandat

Differential Revision: D13560964

fbshipit-source-id: 7014e276c9081649fde6befb114f84d88db39735