Revamp random restarts, improved initialization heuristic for candidate generation
Summary:
This diff introduces 2 major changes:

1. **Adds a new heuristic for generating a set of (q-batch) initial conditions for candidate generation.**
This heuristic performs batch evaluation of the acquisition function to compute function values for all q-batches in supplied samples `X` (ideally qMC samples). It allows to utilize the `covar_module` (more generally, any similarity measure) of a model to compute covariances between points, from which it derives a permutation-invariant metric which is used to encourage "diversity" between points. At the end of the day, it samples from the input candidates according to weights that represent an "exploitation/exploration" tradeoff. This tradeoff is controlled via the `eta_Y` and `eta_sim` options (see docstring).
Computing the permutation invariant metric can be costly for large `q`, so this can be changed to an approximate permutation-invariant metric by limiting the comparison to a random subset of all possible permutations. By setting the `eta_sim` option to zero, similarity in feature space is fully ignored, in which case most of the computations are being avoided and the heuristic is fast.

2. **This diff also changes the way random restart optimization is being done**
It splits optimization from initialization, so that the `b x q x d`-batched set of initial conditions is passed in to `gen_candidates_scipy`. There is a lightweight helper function `get_best_candidates` that extracts the best of the optimized candidates. This is done in order to avoid having "random restart inflation", where we define a new optimization method for each initialization method.

These changes required a number of other changes throughout.

Reviewed By: bletham

Differential Revision: D13552294

fbshipit-source-id: cd1736df4d78e88ed2e28d9101ecc94b40498d83