sequential greedy optimization
Summary:
It's possible that the joint optimization is difficult and that a
sequential greedy optimization would work better for maximizing acquisition
functions.  This adds support for the sequential greedy optimization for
q-batch acquisition functions by leveraging the support for pending
observations.

It also adds a fine-tuning alternative that utilizes the sequential greedy optimization
as an initialization to the joint optimization.

We can potentially improve this later by special-casing the behavior for q=1
batch acquisition functions when closed forms are available.

Reviewed By: Balandat

Differential Revision: D13519985

fbshipit-source-id: 8a135971dd9365e32a1a5118216bbd88e3641d40