Allow reading in parameter_bounds from the module in fit_model
Summary:
In addition to the user_supplied bounds dict for fit_model, the optimization now checks for a `parameter_bounds` dictionary on the model of the same type.
This allows specifying the bounds with the model instead of manually passing them in everywhere. Ideally, this funcitonality would be added upstream
in gpytorch so that bounds can be passed in upon construction (which would also allow defining the bounds on the module to which the parameter belongs,
and hence enable traversing the module tree recursively like we do with parameters etc.)

I will add reasonable default parameter constraints to the standard botorch models in a separate diff.

Reviewed By: sdaulton

Differential Revision: D13823751

fbshipit-source-id: 2fc494ce95cc817674cb577ddb56fc313cb337e2