MultiOutputGP model
Summary:
This is a simple botorch model that is not much more than a container of individual botorch models.

The primary benefits over multi-task GPs (either Kronecker or Hadamard) are:
- this model allows for different hyperapameters / kernels / priors etc. for each outcome
- the number of training examples can be different for the different outcomes

The primary drawbacks are:
- evaluation internally is sequential, so this is potentially quite slow when predicting many outcomes
- correlations between outcomes are not accounted for

Compared to the Multi-Task GPs, subsetting the outputs is straightforward (utilizing the new `from_independent_mvns` class method of the `MultitaskMultivariateNormal` I recently committed upstream).
This allows to train the model for many outputs, and then during acquisition fucntion optimization subset to only those outputs that are utilized in the objective or constraints.

Note that while the number of training examples can vary across outcomes, the `posterior` implementation limits the model to a "block" posterior evaluation
at the same `X`, yielding a (trivial) `MultitaskMultivariateNormal` (this is why MultiTaskGP models are subclassed from MultiOutputGP models...)

Most of the functionality here in fact is inherited from the draft `IndependentMultiOutputGP` in gpytorch.

On the botorch side the API is the same, and things like `fit_model` and all acuqisition function optimization should work out of the box (there will be some changes necessary to the benchmarking code, which currently assumes single-output models throughout).

This diff also removes the `add_observation_noise` function of botorch models, which so far has not really been necessary and would have complicated this diff.

Reviewed By: sdaulton

Differential Revision: D13743492

fbshipit-source-id: 2c60226e04165e91b4ef7ee5c9d4175b61a61c32