Fix noise level in SingleTaskGP if variance is close to constant
Summary: SingleTaskGP in botorch by default estimates the noise level. On the AE side, we currently always have known noise. Rather than have to use the full heteroskedastic model for every fit, this allows a known homoskedastic noise level to be fixed in the SingleTaskGP, and makes that the default on the AE side if the variance in observed noise levels is relatively low.

Reviewed By: Balandat

Differential Revision: D14060091

fbshipit-source-id: e1b34adb6b94a6627cb477397247667b2c4c5527