Support one outcome in botorch multioutput model
Summary: Right now the botorch multioutput model throws an error if you try to make a prediction with only one outcome. This fixes that issue on the botorch side by falling back to a usual MVN, and adjusts the predict on the AE side to return the correct shape in that case.

Reviewed By: Balandat

Differential Revision: D14176427

fbshipit-source-id: ef0ecf57878756f3b0d9c7bb03813bda7296edf6