Properly handle resampling of base samples
Summary:
Previously the default behavior was to always resample (unwantedly).
- This gives samples a `resample` arg that, if True, makes sure that base samples are re-constructed in every  `forward` call.
- This also makes `seed` behavior consistent: If the `seed` kwarg is None, then a random seed is generated. Seed is incremented with each foward call in which the base samples are regenerated.
- Note that this means that even with `seed=None` the samples generated in later evaluations are deterministic conditional on the randomly generated intial seed.

Fixing the base samples, I observe much higher wall times, likely b/c the function is now actually deterministic as well, and the optimizer terminates much later.
That is, because of this issue we have essentially been running L-BFGS-B on a stochastic problem, and so the optimizer often must have just thrown its hands up in the air.
I'm surprised that the closed-loop performance on the Hartmann problems was still very good and competitive with the GPy implementation - this must mean
that in the stochastic setting the optimizer got most of the way there, and that exactly solving the problem was actually not that important in these benchmarks

That said, that need not be the case for akshararai's problem, which *could* explain why closed-loop performance was quite poor.

danielrjiang, I suspect this is related to the high wall times that you have observed with the analytic acquisition functions. We may just need to tweak the scipy default tolerances to make this work better.

Reviewed By: danielrjiang

Differential Revision: D14537073

fbshipit-source-id: 58bc94449c20e77798046d1468117db227d4061a