Meta learning with RGPE
Summary:
Tutorial using a custom gpytorch model (RGPE) to do meta-learning across related tasks (here a toy 1-D function).

One difference from the paper, which I note in the tutorial, is that I refit the kernel hyperparameters for each fold of LOOCV (whereas the paper re-uses the kernel hyperparameters obtained from fitting a GP on all the data points). I can update this, but I think it provides a nice example of how to do the LOOCV in batch mode in gpy/botorch.

Reviewed By: Balandat

Differential Revision: D14552196

fbshipit-source-id: 2d154400ef034cd5e091a65e276ec93a6d88e6ab