Default scipy solver to SLSQP
Summary:
Previously the default was L-BFGS-B if therer were no linear constraints on the parameters.
The test nb shows a 30% speedup of SLSQP over L-BFGS-B (this is rebased on D14644224).
Since SLSQP is more general in its functionality, we should default to to it even if there are no parameter constraints other than bound constraints.

Not sure what the cause for SLSQP being faster is - the algorighm might be better suited to these problems, or maybe the default tolerances are larger.

Reviewed By: danielrjiang

Differential Revision: D14647072

fbshipit-source-id: 504a8e66352ddd980b31d890e96d3a562cac2c11