More reasonable noise priors
Summary:
The prior for the inferred noise of likelihoods was way off and resulted in models with significantly too small noise levels.
This chooses a more reasonable prior. This should be seen as a band aid.
In general, I think we should avoid having overly informative priors in the open source release.
For internal use, we shold run some meta-analysis on canonical use cases to choose reasonable priors for our experiments.

Reviewed By: danielrjiang

Differential Revision: D14838442

fbshipit-source-id: 7a5002ef116ffd5a8f63d28bc9f8a579731e1b15