Use gpytorch constraints for bounds on parameters during model fitting.
Summary:
gpytorch now supports defining constraints on its submodules. This allows to specify parameter constraints where they belong (on the model), and have model fitting deal with this in a generic way.
Note that constraints that have a transform that is not `None` automatically enforces the constraint by using a transform. This can be an issue for quasi 2nd order optimizers though b/c the objective becomes flat when overshooting past the effective constraint in the line search.
Hence not doing the transform and imposing an explicit constraint is preferred. It may also be beneficial to use the transform in conjunction with an explicit bound - will have to evaluate that more.

Reviewed By: bletham

Differential Revision: D14840983

fbshipit-source-id: 6f52ec9eb0b970a692963083125e58df55a46de5