Fix image corruption (#128)
Summary:
Docusaurus runs image compression under the hood, which messes up the botorch logo lockup svg.

It used to be possible to skip this compression but apparently b/c of a bug in docusaurus that's not possible on recent versions: https://github.com/facebook/Docusaurus/issues/1427

We'll have to hack this until that is fixed.
Pull Request resolved: https://github.com/pytorch/botorch/pull/128

Reviewed By: ldworkin

Differential Revision: D15148030

Pulled By: Balandat

fbshipit-source-id: 619eecba6c28c105499a7c9a0368977026736f2b