Make fantasy NEI more efficient
Summary: `model.get_fantasy_model(inputs, targets, kwargs)` creates a copy of `model` where the  new `b x n` input points and `b x n` fantasy targets are concatenated to the (expanded) original `n` inputs points and `n` observed targets, so the GP's inputs are `b x 2n x d` and the targets are `b x 2n` where b is the number of fantasies. What we want for fantasy NEI, is a new batch mode GP (treating each batch (fantasy) independently) using only the fantasies for the training targets (not the observed targets). Using both is not wrong, but is inefficient since we double the number of training points.

Reviewed By: danielrjiang

Differential Revision: D15166758

fbshipit-source-id: 01b2dee2a93550f6bba500f6a60a7cc676b25e4b