Ignore gh-pages branch from within gh-pages in CircleCI
This is kind of wonky, we need to push a CircleCI config to the gh-pages branch in order to
tell CircleCI to skip it, o/w it'll try to run and error out.

Also added a sphinx run prior to website deployment so the website only gets auto-deployed
if there are no sphinx warnings.