Ignore gh-pages branch from within gh-pages in CircleCI (#147)
Summary:
This is kind of wonky, we need to push a CircleCI config to the gh-pages branch in order to tell CircleCI to skip it, o/w it'll try to run and error out.

Also added a sphinx run prior to website deployment so the website only gets auto-deployed if there are no sphinx warnings.

Pull Request resolved: https://github.com/pytorch/botorch/pull/147

Reviewed By: amurthy1

Differential Revision: D15174879

Pulled By: Balandat

fbshipit-source-id: 4307364e9f3e86cd94017ea53749bf4dda4420cf