batch limit in joint optimize
Summary: add an option, "batch_limit", to limit the batch size for parallel optimization and parallel acquisition evaluation when doing joint_optimize and gen_batch_initial_conditions, respectively. Instead, optimization and evaluation will be done sequentially, with each step having a max batch size of batch_limit. In the case of joint_optimize, one might want to do this if the optimization problem is high-dim and doing too many random restarts in parallel causes issues for the optimizer (especially SLSQP). In the case of gen_batch_initial_conditions, if the acquisition function is complicated, evaluating it at a large number of points in batch mode can cause memory issues.

Reviewed By: Balandat

Differential Revision: D15527459

fbshipit-source-id: 07eaf2abf2ba081df4127a679c9eac0ccdd41a1a