botorch.qmc -> botorch.sampling, move MCSamplers there (#172)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/172

The MCSamplers currently live in the acquisition module. There is no real reason for that.
It also causes issues with circular imports when trying to make fantasizing a part of the Model API.

This diff renames `botorch.qmc` to `botorch.sampling` (since this now does more than just qmc), and moves the samplers into this module.

Note that these are breaking changes that need to be surfaced in the release notes.

Differential Revision: D15808512

fbshipit-source-id: 3456459277876eba4f1022ccec686793b9aba145