Condition + fantasize APIs on the Model level (#173)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/173

This diff makes conditioning on new data and fantasizing first-class citizens of the Model API (though not a requirement for implementing new models). Under the hood, fantasizing just wraps sampling from the posterior and conditioning on this fake data.

Further, instead of the `detach_test_caches=False` kwarg in the posterior, you should now use `propagate_grads=True` in when calling `posterior` with the intent of differentiating the output w.r.t. the input features.

Reviewed By: eytan

Differential Revision: D15785439

fbshipit-source-id: 5fbf455faa51927400187ea130323b20cc1ef860