Fit models in SumMarginalLogLikelihood sequentially by default
Currently, `fit_gpytorch_model` solves a single joint optimization problem to fit ModelListGP (and, more generally, any ModuleList model that has a SumMarginalLogLikelihood).
If there are only a few models then this is just fine. If there are a lot of models, each with a lot of parameters, then the dimension of the optimization problem can become quite large, and the optimizer may have a hard time optimizing the joint problem (as it's not exploiting the indenpendence structure).

This changes the default behavior to fit the models sequentially, solving multiple simpler optimization problems instead a single harder one. This can be overriden by passing in `sequential=False` to `fit_gpytorch_model`

Test Plan:
Unit tests