Fit models in SumMarginalLogLikelihood sequentially by default (#183)
Summary:
Currently, `fit_gpytorch_model` solves a single joint optimization problem to fit ModelListGP (and, more generally, any ModuleList model that has a SumMarginalLogLikelihood).

If there are only a few models then this is just fine. If there are a lot of models, each with a lot of parameters, then the dimension of the optimization problem can become quite large, and the optimizer may have a hard time optimizing the joint problem (as it's not exploiting the independence structure).

This PR changes the default behavior to fit the models sequentially, solving multiple simpler optimization problems instead a single harder one. This can be overridden by passing in `sequential=False` to `fit_gpytorch_model`.

Pull Request resolved: https://github.com/pytorch/botorch/pull/183

Test Plan: Unit tests

Reviewed By: sdaulton

Differential Revision: D15980508

Pulled By: Balandat

fbshipit-source-id: 5c8be7c8b8db4ceaed6a8d1738527f788097a6ba