Catch gpytorch numerical issues and return NaN to the optimizer (#184)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/184

scipy's minimize can (sort of, not really) handle NaNs - doing this for the fitting may help us with the robustness issues.

Basically, in `_scipy_objective_and_grad` we catch the "singularity error" in gpytorch and return `NaN` instead.
L-BFGS-B will then terminate with `success=False` and an "abnormal termination in line search" message.
From some simple toy experiments it appears as if the solver isn't smart enough to back off gradually during the line search.
We'll have to hope that this degeneracy only occurs after the optimizer has mostly converged, in which case terminating at the current iterate will not be terrible.

Therefore it may still be necessary to add enforce explicit bounds that (i) avoid numerical issues and (ii) are not too conservative so as to exclude the actual minimum from the feasible set.

Also, the "overstepping" will be dependent on the initial condition, so it would be desirable to re-start the optimization if the optimizer does not report success.

Reviewed By: danielrjiang

Differential Revision: D15977143

fbshipit-source-id: 999d158b0ba2ce310c180b2ac5eaa71baf7430d5