ModelList <-> BatchedModel converters (#187)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/187

Adds converters between `BatchedMultiOutputGPyTorchModel` and `ModelListGP` and the reverse direction (if applicable).

This is useful e.g. for fitting batched multi-output models with a lot of outputs, where jointly fitting the model can result in inferior model fits (due to the size of the resulting optimization problem). See stacked diff.

This currently does **not** support the following:
- `HeteroskedasticSingleTaskGP`
- custom likelihoods for `SingleTaskGP`

Differential Revision: D15982128

fbshipit-source-id: 62450f7bf44546f61cad8a549a46feb2a3d5b755