Sequentially fit batched models using ModelList converter
Summary:
Fitting batched multi-output models with a lot of outputs jointly can result in inferior model fits (due to the size of the resulting optimization problem).

This makes use of the ModelList <-> BatchedModel converter to fit the models corresponding to the different (independent) outputs individually.

Note that this current form may cause issues with custom likelihoods in `SingleTaskGP` models.

Reviewed By: sdaulton

Differential Revision: D16007838

fbshipit-source-id: 0359d991f25c36f8b297ac24827712ccccae0cfd