Sequentially fit batched models using ModelList converter (#189)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/189

Fitting batched multi-output models with a lot of outputs jointly can result in inferior model fits (due to the size of the resulting optimization problem).

This makes use of the ModelList <-> BatchedModel converter to fit the models corresponding to the different (independent) outputs individually.

Note that this current form may cause issues with custom likelihoods in `SingleTaskGP` models.

Reviewed By: sdaulton

Differential Revision: D16007838

fbshipit-source-id: 530d85aa5b17c0d2aa3bc2f184ad3a20e1994c06