Embeding fidelity kernels into SingleTaskGP (#181)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/181

We write SingleTaskMultiFidelityGP as a subclass of SingleTaskGP to deal with GP models that have fidelity parameters. The kernel used in this new GP model comes from Section 5.3 of https://arxiv.org/abs/1903.04703. We also added mean_module and covar_module kwargs to the SingleTaskGP constructor that default to None.

Reviewed By: Balandat

Differential Revision: D15731799

fbshipit-source-id: 0ad6fbed9e26301ff7a08096208c45888e8bdf8d