FixedFeatureAcquisitionFunction wrapper (#219)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/219

Generates a new acquisition function over a lower-dimensional space by wrapping a base acquisition function and inserting fixed values. This is useful e.g. for performing contextual optimization, where a model is fit on a set of features, but then we'd like to optimize conditional on a particular value for the subset of features.

This works in a fashion similar to concatenating in `X_pending`, only that we concatenate the fixed values in to the feature dimension of the input `X`. Hence things are slightly more complicated b/c the order of the columns matters. The convention used here for this is just one way of doing it.

Reviewed By: VilockLi

Differential Revision: D16566353

fbshipit-source-id: 6f0ac20833e858a7c9a7cba37b39c64b4ebe5a48