Add propagate_grads context manager (#221)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/221

Adds a `botorch.settings` module that introduces a contextmanager for the `propagate_grads` setting.
This cleans up the API by removing the `propagate_grads` kwarg from `Model.posterior`.

The new pattern for propagating gradients to the training inputs of a model is the following:
```
    with settings.propagate_grads(True):
        post_X = self.posterior(X, observation_noise=observation_noise)
```

Right now this is very much just a wrapper around GPyTorch's `detach_test_caches` setting, but this allows implementing everything in a model-agnostic fashion.

Reviewed By: sdaulton

Differential Revision: D16583422

fbshipit-source-id: af070580f3016c57f70b726fc416307b772202e7