Make standardize work on the second to last dimension (#263)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/263

Previously, `standardize` worked on the batch (leading) dimension of a tensor.
This is not the typical use case for batched tensors in BoTorch, now that we
require an explicit output dimension.

This changes `standardize` to standardize on `dim=-2` (and just standardize if the tensor happens to be single-dimensional)

Note that while this is a breaking change, it does not look like it does not require updates in other places.

Reviewed By: danielrjiang

Differential Revision: D16965564

fbshipit-source-id: 114050274fd490045e6cd43aeb9743e3fb174a44