Add input scaling check to standard models (#267)
Summary:
Introduces a `settings.validate_input_scaling` flag that when active results in the input data being checked for normalization/standardization.

In this PR this check is on by default, we may want to make it optional if this is too much of a hassle in practice. Right now as used in the models this by default emits warnings and raises errors only on `NaN` values or negative variances.

Addresses https://github.com/pytorch/botorch/issues/208
Pull Request resolved: https://github.com/pytorch/botorch/pull/267

Reviewed By: danielrjiang

Differential Revision: D17398454

Pulled By: Balandat

fbshipit-source-id: 3e00b5820abf2e9b9dbd2b2880287d468d74b54c