Prune X_baseline in qNEI (#287)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/287

Prune points from an input tensor that are unlikely to be the best point.

Given a model, an objective, and an input tensor `X`, this function returns the subset of points in `X` that have some probability of being the best point under the objective. This function uses sampling to estimate the probabilities, the higher the number of points `n` in `X` the higher the number of samples `num_samples` should be to obtain accurate estimates.

Enabled by default on qNEI. This adds some overhead to instantiating the acquisition function, but it's well worth it when repeatedly evaluating. We could also perform this pruning in a lazy fashion, but that seems unnecessary at this point.

Reviewed By: danielrjiang

Differential Revision: D17739861

fbshipit-source-id: 0bf8d606f1bce9a3659621250f43a05486af2d29