Clean up sphinx sources, enforce docs in CI (#294)
Summary:
This PR introduces validation of sphinx coverage in the CI tests. This means that, going forward,

- Every top-level module (e.g., acquisition, models, etc.) should have a corresponding .rst sphix source file in sphinx/source.
- Every single non-package (i.e. py file) module should be included in an .rst file `automodule::` directive.

This is to ensure that Sphinx will automatically include all members from the module in the documentation, and that we have a complete set of API docs.

In the process of making this work, there were also a number of fixes Necessary to docstrings, in particular related to multi-fidelity functionality that was not covered at all up to now. This also renames some files in order to be more consistent with our general conventions.

Pull Request resolved: https://github.com/pytorch/botorch/pull/294

Test Plan: external CI

Reviewed By: sdaulton

Differential Revision: D17953243

Pulled By: Balandat

fbshipit-source-id: 983b0ef0bc1c0e07bf45594fbeb80c102f60355a