Support specifying observation noise explicitly (#256)
Summary:
This is an early draft of adding support for specifying the observation noise in `posterior` and `fantasize`

In addition to using the observation noise from the likelihood by setting `observation_noise=True`, now `observation_noise` can be a tensor. In that case, the provided noise levels are used directly as the observation noise in the posterior predictive (not in performing inference).

The primary use case for this is if we have auxiliary noise models that should not be used as the likelihood during posterior computations (e.g. b/c the model is fitted to already smoothed data), or because we have some dependency of the observation noise on parameters that we may control, e.g. the fidelity of the evaluation/sample size.

This also introduces a `gpt_posterior_settings` context manager that wraps gpytorch settings that are repeatedly called. This helps keep the code cleaner.

Pull Request resolved: https://github.com/pytorch/botorch/pull/256

Test Plan: unit tests

Reviewed By: bkarrer

Differential Revision: D17147235

Pulled By: Balandat

fbshipit-source-id: fc71760b911b2ae260d86ee10830bf035a459875