Add subset_output functionality to (most) models (#324)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/324

In some cases we want to be able to subset models along the output dimension.
For instance, if we fit a multi-output model with a number of metrics, we may want to optimize an acquisition function with an objective that only involves a subset of the
outputs. By subsetting the model prior to that, we can save a lot of compute.

This diff adds a `subset_output` function to the model API. Calling this on a model with a list of indices will return a new model object that is restricted to the desired outputs.

For some models (e.g `AffineDeterministicModel` or `ModelListGP`) the implementation is trivial. For others it's a little more involved but doable.The main challenge is with things like passing in generic covariance modules - we really don't have any way of knowing what dimensions of the respective buffers and parameters we need to subset / rescale in this case.

Reviewed By: sdaulton

Differential Revision: D18668985

fbshipit-source-id: 41479203f23e8bcfa08bbe5f025ed12f0124a091