Allow custom module-to-array handling in fit_gpytorch_scipy
Summary:
fit_gpytorch_scipy uses a pair of functions to convert all of the named parameters in a module into a numpy array. Sometimes we want to also optimize over things that are not named parameters, as in the stacked diff. This makes it easy to do this by allowing one to custom set three things: the function to convert module to array, the function to set moduel params with an array, and the callable given to scipy minimize that computes f and grad at a particular array.

In order to have _scipy_objective_and_grad as a default kwarg I had to move its function definition above that of fit_gpytorch_scipy. I didn't make any changes to that actual function.

Reviewed By: Balandat

Differential Revision: D18630951

fbshipit-source-id: 4bf9df02ba6efe9fd0e54f1814bc190bfb43bf46