Importance sampling for (q)MC acquisition functions
Summary:
This diff implements importance sampling for the `MCSampler` classes, and hooks a basic version of this up into qExpectedImprovement.

If `best_f` is relatively close to the posterior's maximum mean, and if the variance is low, then a lot of the qEI MC objective samples will have zero contribution to the estimate.

This diff allows to inflates the variance of the base samples, record the likelihood ratio for each sample (relative to a `N(0, I)` base sample)  and then applies a weighted mean to the individual sample contributions.

No doubt smarter things could be done for this.

This results in a reduced variance of the estimator. Note that if `best_f` is low, this will instead *increase* the variance of the estimator. Likely the latter is less of a problem since `best_f` will typically close be to the best modeled values in practice. Additional empirical evaluation, in particular of the closed-loop optimization behavior, is needed.

Differential Revision: D18553363

fbshipit-source-id: 104ef8f8a6eee753782b80f2fe813c3c4039bee7