Fix batch-list conversion of SingleTaskMultiFidelityGP (#370)
Summary:
This fixes an issue where fitting a `SingleTaskMultiFidelityGP` resulted in `fit_gpytorch_model` erroring out white it was trying to perform batch-list conversion for fitting.

This fix isn't particularly pretty. We eventually probably want to change this setup to be a method on the model instead, similar to subset_model.
Pull Request resolved: https://github.com/pytorch/botorch/pull/370

Test Plan: unit tests

Differential Revision: D19809842

Pulled By: Balandat

fbshipit-source-id: e034e9e14b522fc00fe663cae8deb3dd4d10e856