Fixing TestLoader local test discovery
Summary: From the difference in runtimes in the test plan (380s this diff vs 15s the base commit), looks like this was really an issue with using `"."` in `discover` for the test runner, and the tests previously didn't run properly on fb infra. They did however run in OSS, where they were run from a proper root dir.

Differential Revision: D19987477

fbshipit-source-id: 7ff445fcd7539f6c8c0e9dfe2b4133c09dd74c75