Fixing TestLoader local test discovery (#376)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/376

From the difference in runtimes in the test plan (380s this diff vs 15s the base commit), looks like this was really an issue with using `"."` in `discover` for the test runner, and the tests previously didn't run properly on fb infra. They did however run in OSS, where they were run from a proper root dir.

Reviewed By: danielrjiang

Differential Revision: D19987477

fbshipit-source-id: c804d077328b1dab29636cc70785cc517faa0604