Better stopping criteria for stochastic optimization (#392)
Summary:
The current `ConvergenceCriterion` assumes that the optimization is using a descent method. That's not true for stochastic optimization, so using this in `gen_candidates_torch` will very often prematurely terminate optimization. This diff adds a basic version of an exponential moving average smoothing of the losses to avoid this.

Since convergence is generally kind of fuzzy in stochastic optimization, this also renames `ConvergenceCriterion` to `StoppingCriterion`. It introduces an abstract base class, and moves this functionality to the `botorch.optim.stopping` submodule.

Pull Request resolved: https://github.com/pytorch/botorch/pull/392

Test Plan: Unit tests

Reviewed By: sdaulton

Differential Revision: D20423879

Pulled By: Balandat

fbshipit-source-id: 8d3519966486e451abd9d0d4a3f81576caf844ec