Run tutorials in CircleCI (#397)
Summary:
Failing tutorials have been a blind spot for a while - this adds a check to CircleCI to run the tutorials against the current changes on each PR, which will avoid this issue.

Right now there are a couple of tutorials failing, I'll fix those in separate PRs and then merge this one in after.

Pull Request resolved: https://github.com/pytorch/botorch/pull/397

Test Plan: CI

Differential Revision: D20457549

Pulled By: Balandat

fbshipit-source-id: 41b597421ed5027c410856c7748bbcd6ae95c5a7