fix cuda tests (#404)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/404

Previously, we did not check whether the cuda tests actually succeeded. This inspects the `TestResult` from each call to `run()` and prints the error/failure message if the test failed.

Some of the code/imports are shuffled around or copied to avoid import errors without a bunch of sys.path hacking.

Reviewed By: Balandat

Differential Revision: D20501133

fbshipit-source-id: b31ee98c96723e60c172e3997c67b8511fd1bedb