Version 0.2.4 (#442)
Summary:
This is a bump mainly to fix a mysterious issue with the python wheel on pypi being broken. Specifically, in that wheel the `botorch/optim/utils.py` file was not included, which resulted in an `ImportError` for many central components of the code. Interestingly, the source dist (built with the same command) did not have this issue.
Pull Request resolved: https://github.com/pytorch/botorch/pull/442

Reviewed By: sdaulton

Differential Revision: D21538808

Pulled By: Balandat

fbshipit-source-id: 4a963d101c28d17d18897584ccadc9f492a0c914